# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AI Agent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /v1/storage/projects/{projectId}/files:
    get:
      operationId: listProjectFiles
      tags: [storage]
      summary: List project files
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFileList"
    post:
      operationId: createProjectFile
      tags: [storage]
      summary: Create a project file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
  /v1/storage/projects/{projectId}/files/{fileId}:
    get:
      operationId: getProjectFile
      tags: [storage]
      summary: Get a project file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
    put:
      operationId: updateProjectFile
      tags: [storage]
      summary: Update a project file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
    delete:
      operationId: deleteProjectFile
      tags: [storage]
      summary: Delete a project file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # --- Project Memory ---
  /v1/memory/projects/{projectId}:
    get:
      operationId: getProjectMemory
      tags: [memory]
      summary: List all memory entries for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Memory list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryList"
    delete:
      operationId: clearProjectMemory
      tags: [memory]
      summary: Clear all memory for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /v1/memory/projects/{projectId}/{key}:
    get:
      operationId: getMemoryEntry
      tags: [memory]
      summary: Get a memory entry by key
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Memory entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: setMemoryEntry
      tags: [memory]
      summary: Set (upsert) a memory entry
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryEntryInput"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
    delete:
      operationId: deleteMemoryEntry
      tags: [memory]
      summary: Delete a memory entry
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: ["object", "null"]
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth schemas
...
    Project:
      type: object
      required: [id, user_id, name, project_type, status, created_at, updated_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        project_type:
          type: string
          enum: [website, bot]
        status:
          type: string
          enum: [draft, active, archived]
        created_at:
          type: string
        updated_at:
          type: string
...
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, archived]
...
    ProjectFile:
      type: object
      required: [id, project_id, path, created_at, updated_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        path:
          type: string
        content:
          type: ["string", "null"]
        mime_type:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    ProjectFileList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProjectFile"
        total:
          type: integer
    ProjectFileInput:
      type: object
      required: [path]
      properties:
        path:
          type: string
          minLength: 1
        content:
          type: string
        mime_type:
          type: string
    ProjectFileUpdate:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        mime_type:
          type: string

    # Memory schemas
    MemoryEntry:
      type: object
      required: [id, project_id, key, value, scope, created_at, updated_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        key:
          type: string
        value: {}
        scope:
          type: string
          enum: [global, session, agent]
        expires_at:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    MemoryList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MemoryEntry"
        total:
          type: integer
    MemoryEntryInput:
      type: object
      required: [value]
      properties:
        value: {}
        scope:
          type: string
          enum: [global, session, agent]
          default: global
        ttl_seconds:
          type: integer