# lib/api-zod/src/generated/types/memoryEntryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Agent Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { MemoryEntryInputScope } from './memoryEntryInputScope';

export interface MemoryEntryInput {
  value: unknown;
  scope?: MemoryEntryInputScope;
  ttl_seconds?: number;
}