# lib/api-zod/src/generated/types/projectFileList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Agent Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { ProjectFile } from './projectFile';

export interface ProjectFileList {
  items: ProjectFile[];
  total: number;
}