# lib/api-zod/src/generated/types/projectFileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Agent Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface ProjectFileUpdate {
  path?: string;
  content?: string;
  mime_type?: string;
}